of: Add of_flat_dt_get_machine_name to device tree library
Add a function to retreive the machine name from the device tree once
the device tree has been unflattened.  This is particularly useful for
multi-flavored virtual machine boards (like the Xen, virt, and ranchu
boards) where it is useful for user space to know which flavor of a
device-tree driven board you are running.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>