arm: Retreive the machine name from the device tree
Instead of using the mdesc descriptive name for the userspace exported
machine name, use the name from the device tree file.  This is
particularly useful for multi-flavored device-tree driven boards like
the ranchu board using the mach-virt board files.

Note that this patch is not intended for newer kernels that rework all
of the mach-virt logic to let a non-board machine be virt by default,
but it is useful in the context of Linux v3.10-based Android kernels.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>