arm: defconfig: Add ranchu_defconfig for Android QEMU emulator
Add a defconfig for the ranchu baord emulated by recent QEMU versions to
serve as a base for the Android emulator.

This is basically a tweaked vexpress_defconfig with ARCH_VIRT, virtio
drivers, and Android drivers enabled plus other bits needed by Android
(like ipfilter).  A pure ARCH_VIRT configuration would be nice, but that
currenly doesn't permit inclusion of AMBA devices like the pl011 used
for the console, and changing the config system on a stable kernel to
allow that is not worth the effort and intrusiveness for a slightly
smaller kernel binary.  Newer versions of Linux will allow for a more
compact kernel/config.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>